+++ src/Monitor/Monitor/Diagnostics/GetAzureRmDiagnosticSettingCommand.cs
@@ -15,6 +15,7 @@
 using Microsoft.Azure.Commands.Insights.OutputClasses;
 using Microsoft.Azure.Management.Monitor;
 using Microsoft.Azure.Management.Monitor.Models;
+using Microsoft.WindowsAzure.Commands.Utilities.Common;
 using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
@@ -24,18 +25,11 @@ namespace Microsoft.Azure.Commands.Insights.Diagnostics
     /// <summary>
     /// Gets the logs and metrics for the resource.
     /// </summary>
-    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DiagnosticSetting"), OutputType(typeof(PSServiceDiagnosticSettings))]
-    public class GetAzureRmDiagnosticSettingCommand : ManagementCmdletBase
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DiagnosticSetting", DefaultParameterSetName = ResourceIdParameterSet), OutputType(typeof(PSServiceDiagnosticSettings))]
+    public class GetAzureRmDiagnosticSettingCommand : DiagnosticSettingCommandBase
     {
 
         #region Parameters declarations
-
-        /// <summary>
-        /// Gets or sets the resourceId parameter of the cmdlet
-        /// </summary>
-        [Parameter(Position = 0, Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The ResourceId")]
-        [ValidateNotNullOrEmpty]
-        public string ResourceId { get; set; }
 
         /// <summary>
         /// Gets or sets the diagnostics setting name parameter of the cmdlet
@@ -48,17 +42,19 @@ namespace Microsoft.Azure.Commands.Insights.Diagnostics
 
         protected override void ProcessRecordInternal()
         {
+            Validate();
+
             IList<PSServiceDiagnosticSettings> output;
             if (string.IsNullOrWhiteSpace(this.Name))
             {
                 // Temporary service name constant provided for backwards compatibility
-                IList<DiagnosticSettingsResource> results = this.MonitorManagementClient.DiagnosticSettings.List(resourceUri: this.ResourceId).Value;
+                IList<DiagnosticSettingsResource> results = this.MonitorManagementClient.DiagnosticSettings.List(resourceUri: GetTargetUri()).Value;
                 output = results.Select(e => new PSServiceDiagnosticSettings(e)).ToList();
             }
             else
             {
                 // Temporary service name constant provided for backwards compatibility
-                DiagnosticSettingsResource result = this.MonitorManagementClient.DiagnosticSettings.Get(resourceUri: this.ResourceId, name: this.Name);
+                DiagnosticSettingsResource result = this.MonitorManagementClient.DiagnosticSettings.Get(resourceUri: GetTargetUri(), name: this.Name);
                 output = new List<PSServiceDiagnosticSettings> { new PSServiceDiagnosticSettings(result) };
             }
             WriteObject(sendToPipeline: output, enumerateCollection: true);