+++ src/Monitor/Monitor/ChangeLog.md
@@ -19,6 +19,12 @@
 -->
 
 ## Upcoming Release
+* Added cmdlet to get diagnostic setting categories for subscription
+    - `Get-AzSubscriptionDiagnosticSettingCategory`
+* Supported subscription diagnostic setting operations with new parameter: SubscriptionId
+    - 'Get-AzDiagnosticSetting'
+    - 'New-AzDiagnosticSetting'
+    - 'Remove-AzDiagnosticSetting'
 
 ## Version 2.4.0
 * Added cmdlets for data collection rules: