+++ src/ServiceFabric/ServiceFabric/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 2.3.0
 * Added parameters `VMImagePublisher`, `VMImageOffer`, `VMImageSku`, `VMImageVersion` to `Add-AzServiceFabricNodeType` to facilitate easy alternate OS image creation for new node type.
 * Added parameter `IsPrimaryNodeType` to `Add-AzServiceFabricNodeType` to be able to create an additional primary node type, for the purpose of transitioning the primary node type to another one in the case of OS migration.
 * `Add-AzServiceFabricNodeType` now correctly copies the LinuxDiagnostic extension. This was previously not working for Linux.
@@ -27,6 +29,25 @@
 * Added cmdlet `Update-AzServiceFabricVmImage` to update the delivered SF runtime package type. This must be changed when migrating from Ubuntu 16 to 18.
 * Added cmdlet `Update-AzServiceFabricNodeType` to update the properties of a cluster node type. For now this is solely used to update whether the node type is primary via bool parameter `-IsPrimaryNodeType $false`.
 * `Update-AzServiceFabricReliability` is now able to update reliability level when the cluster has more than one primary node type. To do this, the name of the node type is supplied via the new -NodeType parameter.
+
+* Added new cmdlets for managed applications:
+    - `New-AzServiceFabricManagedClusterApplication`
+    - `Get-AzServiceFabricManagedClusterApplication`
+    - `Set-AzServiceFabricManagedClusterApplication`
+    - `Remove-AzServiceFabricManagedClusterApplication`
+    - `New-AzServiceFabricManagedClusterApplicationType`
+    - `Get-AzServiceFabricManagedClusterApplicationType`
+    - `Set-AzServiceFabricManagedClusterApplicationType`
+    - `Remove-AzServiceFabricManagedClusterApplicationType`
+    - `New-AzServiceFabricManagedClusterApplicationTypeVersion`
+    - `Get-AzServiceFabricManagedClusterApplicationTypeVersion`
+    - `Set-AzServiceFabricManagedClusterApplicationTypeVersion`
+    - `Remove-AzServiceFabricManagedClusterApplicationTypeVersion`
+    - `New-AzServiceFabricManagedClusterService`
+    - `Get-AzServiceFabricManagedClusterService`
+    - `Set-AzServiceFabricManagedClusterService`
+    - `Remove-AzServiceFabricManagedClusterService`
+* Upgraded Managed Cluster commands to use Service Fabric Managed Cluster SDK version 1.0.0-beta.1 which uses service fabric resource provider api-version 2021-01-01-preview.
 
 ## Version 2.2.2
 * Fixed `Add-AzServiceFabricNodeType`. Added node type to service fabric cluster before creating virtual machine scale set.